requirements.txt
Django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
django-cors-headers==4.3.1
django-health-check==3.18.1
psycopg2-binary==2.9.9
django-redis==5.4.0
gunicorn==21.2.0
